requirements.txt
cachetools==5.0.0
certifi==2021.10.8
charset-normalizer==2.0.12
click==8.0.4
colorama==0.4.4
dnspython==2.2.1
dominate==2.6.0
email-validator==1.1.3
Flask==2.0.3
Flask-Bootstrap==3.3.7.1
Flask-Bootstrap4==4.0.2
flask-bootstrap5==0.1.dev1
Flask-SQLAlchemy==2.5.1
Flask-Uploads==0.2.1
Flask-WTF==1.0.0
google-api-core==2.7.1
google-auth==2.6.2
google-cloud-vision==2.7.1
googleapis-common-protos==1.56.0
greenlet==1.1.2
grpcio==1.44.0
grpcio-status==1.44.0
gunicorn==20.1.0
idna==3.3
itsdangerous==2.1.1
Jinja2==3.0.3
joblib==1.1.0
MarkupSafe==2.1.1
numpy==1.22.3
opencv-python==4.5.5.64
pandas==1.4.1
Pillow==9.0.1
proto-plus==1.20.3
protobuf==3.19.4
pyasn1==0.4.8
pyasn1-modules==0.2.8
python-dateutil==2.8.2
pytz==2021.3
regex==2022.3.15
requests==2.27.1
rsa==4.8
scikit-learn==1.0.2
scipy==1.8.0
six==1.16.0
SQLAlchemy==1.4.34
threadpoolctl==3.1.0
tqdm==4.63.0
urllib3==1.26.9
visitor==0.1.3
Werkzeug==2.0.3
WTForms==3.0.1
